Move raster cache state to map store
